Move expanded.py into output_formatter package.
