+++ mycli/output_formatter/expanded.py
